Move storage-contract test bases into `server-test-fixtures`
The storage-contract test suites (`AggregateStorageTest`, `InboxStorageTest`,
`DeliveryTest`, `CatchUpTest`, `ShardedWorkRegistryTest`,
`RecordStorageDelegateTest`, `DefaultEventStoreTest`,
`AggregateHistoryTruncationTest`) and their parents (`AbstractStorageTest`,
`AbstractDeliveryTest`) are reused by downstream storage modules (e.g.
`gcloud-java`, `jdbc-storage`), which subclass them and swap in a
backend-specific `StorageFactory`. They lived in `server/src/test`, so they were
not part of the published `server-test-fixtures` variant and could not be
compiled against the published artifacts.

Relocate the 10 base/parent classes to `server/src/testFixtures` (the test proto
and `given` helpers they depend on are already there) so they ship in
`server-test-fixtures`. The six formerly-concrete suites become `abstract`, and
`AbstractDeliveryTest` becomes `public`, so they are subclassable cross-module
and are not executed directly.

To preserve the in-memory coverage those concrete suites provided in this module,
add six thin `InMemory*Test` runners under `server/src/test` (Gradle's `test`
task discovers only the `test` source set, not `testFixtures`). They follow the
existing `InMemoryShardedWorkRegistryTest` /
`InMemoryAggregateHistoryTruncationTest` pattern and run the inherited contracts
against the default in-memory `StorageFactory`.
