+++ server/src/testFixtures/java/io/spine/server/aggregate/AggregateHistoryTruncationTest.java
