+++ server/src/testFixtures/java/io/spine/server/delivery/AbstractDeliveryTest.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2026, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.delivery;
+
+import com.google.protobuf.util.Durations;
+import io.spine.environment.Tests;
+import io.spine.server.ServerEnvironment;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+
+/**
+ * An abstract base for the tests performing actual delivery.
+ */
+public abstract class AbstractDeliveryTest {
+
+    private Delivery originalDelivery;
+
+    @BeforeEach
+    public void setUp() {
+        this.originalDelivery = ServerEnvironment.instance()
+                                                 .delivery();
+    }
+
+    @AfterEach
+    public void tearDown() {
+        ServerEnvironment.when(Tests.class)
+                         .use(originalDelivery);
+    }
+
+    static void changeShardCountTo(int shards) {
+        var newDelivery = Delivery.localWithShardsAndWindow(shards, Durations.ZERO);
+        ServerEnvironment.when(Tests.class)
+                         .use(newDelivery);
+    }
+}
