+++ server/src/testFixtures/java/io/spine/server/delivery/DeliveryTest.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2026, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -75,7 +75,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 @SlowTest
 @DisplayName("Delivery of messages to entities should deliver those via")
 @SuppressWarnings("WeakerAccess")   // Exposed for libraries, wishing to run these tests.
-public class DeliveryTest extends AbstractDeliveryTest {
+public abstract class DeliveryTest extends AbstractDeliveryTest {
 
     @Test
     @DisplayName("a single shard to a single target in a multi-threaded env")
@@ -481,7 +481,6 @@ public class DeliveryTest extends AbstractDeliveryTest {
         private ImmutableList<DeliveryStats> stats() {
             return ImmutableList.copyOf(allStats);
         }
-
 
         @Override
         public FailedPickUp.Action onShardPickUpFailure(RuntimeFailure failure) {
