+++ server/src/testFixtures/java/io/spine/server/delivery/InboxStorageTest.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2026, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -65,7 +65,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * An abstract base for tests of {@link InboxStorage} implementations.
  */
 @DisplayName("`InboxStorage` should")
-public class InboxStorageTest
+public abstract class InboxStorageTest
         extends AbstractStorageTest<InboxMessageId, InboxMessage, InboxStorage> {
 
     private static final String TARGET_ID = "the-storage-calc";
