+++ server/src/testFixtures/java/io/spine/server/storage/AbstractStorageTest.java
