+++ server/src/testFixtures/java/io/spine/server/storage/RecordStorageDelegateTest.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2026, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -90,7 +90,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
  * Therefore all tests will run against the desired storage engine.
  */
 @DisplayName("A `RecordStorageDelegate` descendant should")
-public class RecordStorageDelegateTest
+public abstract class RecordStorageDelegateTest
         extends AbstractStorageTest<StgProjectId, StgProject, StgProjectStorage> {
 
     @Override
